pyproject.toml
[project]
name = "llm-finetune"
version = "0.1.0"
description = "CLI tool for LLM finetuning with Ollama integration"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "Tim" }]
keywords = ["llm", "finetuning", "ollama", "unsloth", "lora"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "typer>=0.9.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
train = [
    "unsloth",
    "trl>=0.7",
    "transformers>=4.40",
    "datasets>=2.0",
    "torch>=2.0",
    "accelerate>=0.30",
    "bitsandbytes>=0.43",
    "peft>=0.10",
]
dev = [
    "pre-commit",
    "black",
    "flake8",
    "isort",
    "mypy",
    "pytest",
    "pytest-cov",
    "types-PyYAML",
]
all = ["llm-finetune[train,dev]"]

[project.scripts]
finetune = "finetune.cli.main:app"

[project.urls]
Homepage = "https://github.com/tim/local-ai-interface"
Repository = "https://github.com/tim/local-ai-interface"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["finetune"]

# Tool configurations

[tool.black]
line-length = 88
target-version = ["py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
    \.eggs
    | \.git
    | \.hatch
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | checkpoints
    | exports
)/
'''

[tool.isort]
profile = "black"
line_length = 88
skip_gitignore = true
skip = ["checkpoints", "exports", ".venv"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]
exclude = ["checkpoints", "exports", "tests"]

[[tool.mypy.overrides]]
module = [
    "unsloth.*",
    "trl.*",
    "transformers.*",
    "datasets.*",
    "torch.*",
    "accelerate.*",
    "bitsandbytes.*",
    "peft.*",
]
ignore_missing_imports = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=finetune --cov-report=term-missing --cov-report=html"
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["finetune"]
omit = ["finetune/__main__.py"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
fail_under = 70
